requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
cffi==1.15.1
charset-normalizer==3.1.0
discord-interactions==0.4.0
discord-menu==0.16.16
discord-py-interactions==5.7.0
discord-typings==0.6.0
discord-ui==5.1.6
discord.py==2.3.1
Django==4.2.2
emoji==2.6.0
frozenlist==1.3.3
game==0.0.5
idna==3.4
multidict==6.0.4
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
yarl==1.9.2
